requirements.txt
pytorch==1.5.1
numpy>=1.17.3
scikit-learn==0.23.2
pandas>=1.0.4
androguard==3.3.5
absl-py==0.8.1
# python-magic-bin==0.4.14
seaborn
Pillow
pyelftools
capstone
python-magic
pyyaml
multiprocessing-logging
argparse==1.4.0
backcall==0.1.0
block-bootstrap-pytorch==0.1.5
bootstrap-pytorch==0.0.13
click==7.0
cycler==0.10.0
decorator==4.4.0
h5py==2.10.0
ipdb==0.12.2
ipython==7.8.0
ipython-genutils==0.2.0
jedi==0.15.1
kiwisolver==1.1.0
matplotlib==3.1.1
munch==2.3.2
nltk==3.5
opencv-python==4.1.1.26
pandas==0.25.1
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
pillow==6.2.0
plotly==4.1.1
pretrainedmodels==0.7.4
prompt-toolkit==2.0.9
protobuf==3.9.2
ptyprocess==0.6.0
pyemd==0.5.1
pygments==2.4.2
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.2
pyyaml==5.1.2
retrying==1.3.3
scipy==1.3.1
seaborn==0.9.0
skipthoughts==0.0.1
tabulate==0.8.5
tensorboardx==1.8
torchvision==0.4.0
tqdm==4.36.1
traitlets==4.3.2
transformers==4.14.1
wcwidth==0.1.7